add: validation for Trigger component
