+++ scripts/validate.ts
@@ -1,4 +1,5 @@
 import { LocalFileSystem } from './utils/local'
+import { getTriggerComponent, assertValidTriggerComponent } from './utils/types'
 
 async function main() {
   const local = new LocalFileSystem('./packs')
@@ -8,6 +9,8 @@ async function main() {
     const assets = await local.getAssets(assetsPath)
     for (const asset of assets) {
       console.log(asset.name, '✅')
+      const triggerComponent = getTriggerComponent(asset)
+      if (triggerComponent) assertValidTriggerComponent(asset.name, triggerComponent)
     }
     console.log(assetPack.name, '✅')
   }
